+++ yolort/v5/models/__init__.py
@@ -5,12 +5,13 @@ from .common import (
     focus_transform, space_to_depth,
 )
 from .yolo import Detect, Model
-from .experimental import attempt_load
+from .experimental import Ensemble
 
 __all__ = [
     'Conv', 'Bottleneck', 'SPP', 'SPPF', 'DWConv', 'Focus',
     'BottleneckCSP', 'C3', 'Concat', 'GhostConv',
     'GhostBottleneck', 'AutoShape', 'Contract', 'Expand',
-    'focus_transform', 'space_to_depth', 'attempt_load',
+    'focus_transform', 'space_to_depth',
     'Detect', 'Model',
+    'Ensemble',
 ]
