+++ yolort/v5/models/experimental.py
@@ -7,8 +7,7 @@ import numpy as np
 import torch
 from torch import nn
 
-from yolort.ultralytics.utils import attempt_download
-from .common import Conv
+from . import Conv
 
 
 class CrossConv(nn.Module):
@@ -83,33 +82,3 @@ class Ensemble(nn.ModuleList):
 
         y = torch.cat(y, 1)  # nms ensemble
         return y, None  # inference, train output
-
-
-def attempt_load(weights, map_location=None, inplace=True, fuse=True):
-    from .yolo import Detect, Model
-
-    # Loads an ensemble of models weights=[a,b,c] or a single model weights=[a] or weights=a
-    model = Ensemble()
-    for w in weights if isinstance(weights, list) else [weights]:
-        ckpt = torch.load(attempt_download(w), map_location=map_location)  # load
-        if fuse:
-            model.append(ckpt['ema' if ckpt.get('ema') else 'model'].float().fuse().eval())  # FP32 model
-        else:
-            model.append(ckpt['ema' if ckpt.get('ema') else 'model'].float().eval())  # without layer fuse
-
-    # Compatibility updates
-    for m in model.modules():
-        if type(m) in [nn.Hardswish, nn.LeakyReLU, nn.ReLU, nn.ReLU6, nn.SiLU, Detect, Model]:
-            m.inplace = inplace  # pytorch 1.7.0 compatibility
-        elif type(m) is Conv:
-            m._non_persistent_buffers_set = set()  # pytorch 1.6.0 compatibility
-
-    if len(model) == 1:
-        return model[-1]  # return model
-    else:
-        print(f'Ensemble created with {weights}\n')
-        for k in ['names']:
-            setattr(model, k, getattr(model[-1], k))
-        # max stride
-        model.stride = model[torch.argmax(torch.tensor([m.stride.max() for m in model])).int()].stride
-        return model  # return ensemble
